update test script
