+++ mainnet_contract/package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@nomiclabs/hardhat-ethers": "^2.2.1",
     "@nomiclabs/hardhat-waffle": "^2.0.6",
+    "chai": "^4.3.7",
     "dotenv": "^16.1.4",
     "ethereum-waffle": "^3.2.0",
     "hardhat": "^2.12.2",
