+++ mainnet_contract/test/subnet.js
@@ -1,434 +1,66 @@
-const { forContractInstance } =  require("@truffle/decoder");
-const RLP = require("rlp");
-const util = require("@ethereumjs/util");
-const secp256k1 = require("secp256k1");
-const HeaderReader = artifacts.require("HeaderReader");
-const Subnet = artifacts.require("Subnet");
-
-const num2Arr = (n) => {
-  if (!n) return new Uint8Array(0)
-  const a = []
-  a.unshift(n & 255)
-  while (n >= 256) {
-    n = n >>> 8
-    a.unshift(n & 255)
-  }
-  return new Uint8Array(a);
-}
+const { expect } = require("chai");
+const { ethers } = require("hardhat");
 
 const hex2Arr = (hexString) => {
   if (hexString.length % 2 !== 0) {
-      throw "Must have an even number of hex digits to convert to bytes";
+    throw "Must have an even number of hex digits to convert to bytes";
   }
   var numBytes = hexString.length / 2;
   var byteArray = new Uint8Array(numBytes);
-  for (var i=0; i<numBytes; i++) {
-      byteArray[i] = parseInt(hexString.substr(i*2, 2), 16);
+  for (var i = 0; i < numBytes; i++) {
+    byteArray[i] = parseInt(hexString.substr(i * 2, 2), 16);
   }
   return byteArray;
+};
+function blockToHash(blockEncoded) {
+  return ethers.utils
+    .keccak256(Buffer.from(hex2Arr(blockEncoded.slice(2))))
+    .toString("hex");
 }
 
-const composeAndSignBlock = (number, round_num, prn, parent_hash, validators, threshold, current, next) => {
+describe("Subnet", () => {
+  let subnetTest;
+  beforeEach("deploy SubnetTest", async () => {
+    const headerReaderFactory = await ethers.getContractFactory("HeaderReader");
 
-  const version = new Uint8Array([2]);
-  const voteForSignHash = web3.utils.sha3(Buffer.from(
-    RLP.encode([
-      [parent_hash, prn, number-1],
-      0
-    ])
-  ));
-
-  var raw_sigs = []
-  for (let i = 0; i < threshold; i++) {
-    raw_sigs.push(
-      secp256k1.ecdsaSign(
-        hex2Arr(voteForSignHash.substring(2)),
-        hex2Arr(validators[i].privateKey.substring(2))
-    ));
-  }
-  var sigs = raw_sigs.map(x => {
-    var res = new Uint8Array(65);
-    res.set(x.signature, 0);
-    res.set([x.recid], 64);
-    return "0x"+Buffer.from(res).toString("hex");
-  });
-
-
-  var block = {
-    "parent_hash": parent_hash,
-    "uncle_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "coinbase": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "root": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "txHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "receiptAddress": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "bloom": new Uint8Array(256),
-    "difficulty": 0,
-    "number": number,
-    "gasLimit": 0,
-    "gasUsed": 0,
-    "time": 0,
-    "extra": new Uint8Array([...version, ...RLP.encode([
-      round_num,
-      [
-        [parent_hash, prn, number-1],
-        sigs,
-        0
-      ]
-    ])]),
-    "mixHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "nonce": new Uint8Array(8),
-    "validator": new Uint8Array(8),
-    "validators": current,
-    "nextValidators": next,
-    "penalties": new Uint8Array(8)
-  }
-
-  var block_encoded = Buffer.from(RLP.encode([
-    util.toBuffer(parent_hash),
-    util.zeros(32),
-    util.zeros(32),
-    util.zeros(32),
-    util.zeros(32),
-    util.zeros(32),
-    new Uint8Array(256),
-    util.bigIntToUnpaddedBuffer(0),
-    util.bigIntToUnpaddedBuffer(number),
-    util.bigIntToUnpaddedBuffer(0),
-    util.bigIntToUnpaddedBuffer(0),
-    util.bigIntToUnpaddedBuffer(0),
-    new Uint8Array([...version, ...RLP.encode([
-      round_num,
-      [
-        [parent_hash, prn, number-1],
-        sigs,
-        0
-      ]
-    ])]),
-    util.zeros(32),
-    new Uint8Array(8),
-    new Uint8Array(8),
-    current, 
-    next,
-    new Uint8Array(8),
-  ]));
-
-  var block_hash = web3.utils.sha3(block_encoded).toString("hex");
-  return [block, "0x"+block_encoded.toString("hex"), block_hash];
-}
-
-const createValidators = (num_val) => {
-  const validators = [];
-  for (let i = 0; i < num_val; i++) {
-    validators.push(web3.eth.accounts.create());
-  }
-  return validators;
-}
-
-contract("Subnet Test", async accounts => {
-
-  beforeEach(async () => {
-    this.validators = [];
-    this.validators_addr = [];
-    for (let i = 0; i < 3; i++) {
-      this.validators.push(web3.eth.accounts.create());
-      this.validators_addr.push(this.validators.at(-1).address);
-    }
-    const version = new Uint8Array([2]);
-    const voteForSignHash = web3.utils.sha3(Buffer.from(
-      RLP.encode([
-        ["0x0000000000000000000000000000000000000000000000000000000000000000", 0, 0],
-        0
-      ])
-    ));
-    const raw_sigs = [];
-    for (let i = 0; i < 3; i++) {
-      raw_sigs.push(
-        secp256k1.ecdsaSign(
-          hex2Arr(voteForSignHash.substring(2)),
-          hex2Arr(this.validators[i].privateKey.substring(2))
-        )
-      );
-    }
-    const sigs = raw_sigs.map(x => {
-      var res = new Uint8Array(65);
-      res.set(x.signature, 0);
-      res.set([x.recid], 64);
-      return "0x"+Buffer.from(res).toString("hex");
+    const headerReader = await headerReaderFactory.deploy();
+    await headerReader.deployed();
+    const factory = await ethers.getContractFactory("Subnet", {
+      libraries: {
+        HeaderReader: headerReader.address,
+      },
     });
-
-    this.genesis_block = {
-      "parent_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "uncle_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "coinbase": "0x0000000000000000000000000000000000000000",
-      "root": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "txHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "receiptAddress": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "bloom": new Uint8Array(256),
-      "difficulty": 0,
-      "number": 0,
-      "gasLimit": 0,
-      "gasUsed": 0,
-      "time": 0,
-      "extra": new Uint8Array([...version, ...RLP.encode([
-        0,
-        [
-          ["0x0000000000000000000000000000000000000000000000000000000000000000", 0, 0],
-          sigs,
-          0
-        ]
-      ])]),
-      "mixHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "nonce": new Uint8Array(8),
-      "validator": new Uint8Array(8),
-      "validators": [],
-      "nextValidators": [],
-      "penalties": new Uint8Array(8)
-    }
-    
-    this.genesis_hash = web3.utils.sha3(Buffer.from(
-      RLP.encode([
-        util.zeros(32),
-        util.zeros(32),
-        util.zeros(32),
-        util.zeros(32),
-        util.zeros(32),
-        util.zeros(32),
-        new Uint8Array(256),
-        util.bigIntToUnpaddedBuffer(0),
-        util.bigIntToUnpaddedBuffer(0),
-        util.bigIntToUnpaddedBuffer(0),
-        util.bigIntToUnpaddedBuffer(0),
-        util.bigIntToUnpaddedBuffer(0),
-        new Uint8Array([...version, ...RLP.encode([
-          0,
-          [
-            ["0x0000000000000000000000000000000000000000000000000000000000000000", 0, 0],
-            sigs,
-            0
-          ]
-        ])]),
-        util.zeros(32),
-        new Uint8Array(8),
-        new Uint8Array(8),
-        [], [],
-        new Uint8Array(8),
-    ])));
-
-    this.genesis_encoded = "0x"+Buffer.from(RLP.encode([
-      util.zeros(32),
-      util.zeros(32),
-      util.zeros(32),
-      util.zeros(32),
-      util.zeros(32),
-      util.zeros(32),
-      new Uint8Array(256),
-      util.bigIntToUnpaddedBuffer(0),
-      util.bigIntToUnpaddedBuffer(0),
-      util.bigIntToUnpaddedBuffer(0),
-      util.bigIntToUnpaddedBuffer(0),
-      util.bigIntToUnpaddedBuffer(0),
-      new Uint8Array([...version, ...RLP.encode([
-        0,
-        [
-          ["0x0000000000000000000000000000000000000000000000000000000000000000", 0, 0],
-          sigs,
-          0
-        ]
-      ])]),
-      util.zeros(32),
-      new Uint8Array(8),
-      new Uint8Array(8),
-      [], [],
-      new Uint8Array(8),
-    ])).toString("hex");
-    
-    var [block1, block1_encoded, block1_hash] = composeAndSignBlock(1, 1, 0, this.genesis_hash, this.validators, 2, [], []);
-    this.block1 = block1;
-    this.block1_encoded = block1_encoded;
-    this.block1_hash = block1_hash;
-
-    this.lib = await HeaderReader.new();
-    Subnet.link("HeaderReader", this.lib.address);
-    this.subnet = await Subnet.new(
-      this.validators_addr,
-      this.genesis_encoded,
-      this.block1_encoded,
-      5,
-      10,
-      {"from": accounts[0]}
+    subnetTest = await factory.deploy(
+      [
+        "0x888c073313b36cf03CF1f739f39443551Ff12bbE",
+        "0x5058dfE24Ef6b537b5bC47116A45F0428DA182fA",
+        "0xefEA93e384a6ccAaf28E33790a2D1b2625BF964d",
+      ],
+      "0x7fdd36f7a6b4d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d346b4d5d71ce1d7bc75e73be5deda69bf396f9ebb6fa71c778d5a777d76e39d5bf78f1aef8d777f46b5e367dde34778f77e3bf78d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d346b4d7b75fd77e5ae1f69dd7ddf6f777b677cf79d356def1ae7769eefadb6df5d1b6dae7a6fcd1bedee7b6fbe7cd1c79bf7d6b4e7a7bcd5fd7bd5b71ce796ba7dff37e397baf767347fce9ee5be3c7b4d5bf7de9c69d734d35eb6d9f6f97b7eb76f8db56b4e7a7bcd5fd7bd5b71ce796ba7dff37e397baf767347fce9ee5be3c7b4d5bf7de9c69d734d35eb6d9f6f97b7eb76f8db56fdd35d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d35f34f37e3b6fbeb4f34f38eb8e76d777776fcf5dd34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34e74e7c75f7b6e1e7fa6f9dfb6f96dce3bd75e9ae397f4e36f1d6b5f367daf3cf1cd3bdf7d776f7e9c7f4ddc7f57fbdfd7f7f78e37e79d5f7f5d9b6de79f79af777b7f386ba71c69a7f6f1edf7efdd1ad9dd5bdbadb96dff7ae1dd34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d346b4d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34f3cd34d34d34d34d34d34d34d34f34734734f3",
+      "0x7fdd367bb6b4f36737eb4e5b79dd1cdb871fd376b8e5b6f5e3a77cf1eeb66baefbe36735db67bce7af3ae39e5edb6f367f6efdeb5db96b4d5d71ce1d7bc75e73be5deda69bf396f9ebb6fa71c778d5a777d76e39d5bf78f1aef8d777f46b5e367dde34778f77e3bf78f3cf1cd3bdf7d776f7e9c7f4ddc7f57fbdfd7f7f78e37e79d5f7f5d9b6de6b473575cdbcdf8d7b7f6d3d77ad1ad5be1af3ce1de1d6bbd79efcd7cd7b6f96bcddaf5ceb8d9addedde7fdf74d5ad1fdfc6b4d5beb7f1ed1c77ae5cf76d9d6f96bbf1ff76eda79ae7dd1d7367f4e5b7dde1ff3bd34f77f7a776ef4df473a6b8774dbd6b4dfbebbef6d3a73ce5dedb7fdd396b67f6f5df77f3473ae5cf1de9fef6df56f57396b7e9c6fadbc6dce5c71cd1f7b87746fdd35d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d35d35f38d7dd3c6f5d34f36eb7d3cf38eb8ebc75addc69ad367bcd357ba7b76b4f36737eb4e5b79dd1cdb871fd376b8e5b6f5e3a77cf1eeb66baefbe36735db67bce7af3ae39e5edb6f367f6efdeb5db9f34f34734f346b4d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34d34f3cd34d34d34d34d34d34d34d346fce35d9cd3ae38f3bddddf76b4d38e77e3c6dadb8eddd3cddd6dfdb669ef3a71ddf9e34dda73a7f471d7776dc7f673bd5ad9addb6bdefdf5bf7dd5b7f97b579a6f6df9d7cd3d7bb7dcddbe1a779e3ae746dcedf71d6bd7377f9e36d9cd75f75e1af3bd347fcddff78e74e7c75f7b6e1e7fa6f9dfb6f96dce3bd75e9ae397f4e36f1d6b5f367daf78f3cf1cd3bdf7d776f7e9c7f4ddc7f57fbdfd7f7f78e37e79d5f7f5d9b6def7879f79af777b7f386ba71c69a7f6f1edf7efdd1ad9dd5bdbadb96dff7ae1d7fcddff78e74e7c75f7b6e1e7fa6f9dfb6f96dce3bd75e9ae397f4e36f1d6b5f367daf78f3cf1cd3bdf7d776f7e9c7f4ddc7f57fbdfd7f7f78e37e79d5f7f5d9b6def7879f79af777b7f386ba71c69a7f6f1edf7efdd1ad9dd5bdbadb96dff7ae1df34",
+      450,
+      900
     );
-    this.decoder = await forContractInstance(this.subnet);
-  })
-
-  it("Running Setup", async() => {
-    const is_master = await this.subnet.isMaster(accounts[0]);
-    const validators = await this.subnet.getCurrentValidators();
-    assert.equal(is_master, true);
-    assert.deepEqual(validators[0], this.validators_addr);
-    assert.equal(validators[1], 2);
+    await subnetTest.deployed();
   });
+  describe("Run Test", () => {
+    it("Receive New Header", async () => {
+      const block2Encoded =
+        "0xF902F1A0B09A86862B9BBF90E57A81C7F524D0C71E886F716E204C81FD84266B3B779705A01DCC4DE8DEC75D7AAB85B567B6CCD41AD312451B948A7413F0A142FD40D49347945058DFE24EF6B537B5BC47116A45F0428DA182FAA02782BC175424ED837DFEA39BCEF34A75C0BAAACA3083ABAFAD631C46D10A9FF1A0D4E0EFE3CBA3CF72038E7E7F04C62D13425706DD8ED0EAE692ADD3E347E4B5CFA037677206C85D7BF905A2F29D9380C65C8D6F7231B1C5A36CB628BC5CCC0FE4D0B90100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000102841908B1008263088464343A46B8B302F8B003F8ADE3A0B09A86862B9BBF90E57A81C7F524D0C71E886F716E204C81FD84266B3B7797050201F886B8415DCD791A9C53F145F50AF2FCF46E9E4DB60145B484DB8A2DD6C7431CCDF486133A0BF1BE7F79763B6FD91F06BA9A4C3FAAF1B7CF3236D2388F56785E09C460C300B841F1325D2A58ABA54FAF7E707CC1865E8592315ECCE6F281EC43E71C0C36670D0F13CDCA2BC454C555DB9C1F4447850E21D2F005326F33CE1AFE3F984C3E2EE2980180A00000000000000000000000000000000000000000000000000000000000000000880000000000000000B841F0105FD7F3C8D60EA70995D41C7FF3AC843D4B5901B1C61028CDF59B2F2134F9028914EE0C9A6ED47B57F18A3354FE6673A3D07E4DCE71BF64F30A41ABD6D08701C0C080";
+      const block3Encoded =
+        "0xF902F1A02C4265C60FAA7BE948C2BCF897528B5BE49EE255DC63A59AC485A27A791CE455A01DCC4DE8DEC75D7AAB85B567B6CCD41AD312451B948A7413F0A142FD40D4934794888C073313B36CF03CF1F739F39443551FF12BBEA09A6970490B6842D223465AD201E0D51ADF93FFBC7639947679081897168E63CCA0D3B5C496169AE4BF020007192C85041AD9A67DFEF848CB0BF0122B049E01A563A037677206C85D7BF905A2F29D9380C65C8D6F7231B1C5A36CB628BC5CCC0FE4D0B90100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000103841908B1008263088464343A48B8B302F8B004F8ADE3A02C4265C60FAA7BE948C2BCF897528B5BE49EE255DC63A59AC485A27A791CE4550302F886B8413418A10D28A161F27C012288E8894BA59EA03BE35A31AB0773A208D504FE67716BEB8849E3F527E1E489FE5DD28D3B244A9EF61F97D4727A66FEC5B709DF201A01B84166975E377B835A5B5D6997B89595381DF74B908B36DBA755E79CAD2FA805A93E3B45D7C4675139C639845357E80064B56E5FB6BBB9C286C42A6D275A9EC034990180A00000000000000000000000000000000000000000000000000000000000000000880000000000000000B841C19275F72B8DD15D614BAE2FDF603CD04263529D8B1E52B383E048FE3BD0764B22247DE4CEE6FF17443F30F8B39C48E6DEEB21AA700AD4CE99080A868FE8C77B01C0C080";
+      await subnetTest.receiveHeader([block2Encoded, block3Encoded]);
 
-  it("Add Master", async() => {
-    await this.subnet.addMaster(accounts[1]);
-    var is_master = await this.subnet.isMaster(accounts[1]);
-    assert.equal(is_master, true);
-    await this.subnet.removeMaster(accounts[1]);
-    is_master = await this.subnet.isMaster(accounts[1]);
-    assert.equal(is_master, false);
+      const block2Hash = blockToHash(block2Encoded);
+      const block3Hash = blockToHash(block3Encoded);
+      const block2Resp = await this.subnetTest.getHeader(block2Hash);
+      const block3Resp = await this.subnetTest.getHeader(block3Hash);
+      const latestBlocks = await this.subnetTest.getLatestBlocks();
+
+      expect(block2Resp[4]).to.eq(true);
+      expect(block3Resp[4]).to.eq(true);
+      expect(latestBlocks["0"][0]).to.eq(block3Hash);
+      expect(latestBlocks["1"][0]).to.eq(block2Hash);
+    });
   });
-
-  it("Receive New Header", async() => {
-
-    var [block2, block2_encoded, block2_hash] = composeAndSignBlock(2, 2, 1, this.block1_hash, this.validators, 2, [], []);
-
-    await this.subnet.receiveHeader([block2_encoded]);
-
-    const block2_resp = await this.subnet.getHeader(block2_hash);
-    assert.equal(block2_resp[0], this.block1_hash);
-    assert.equal(block2_resp[1], 2);
-    assert.equal(block2_resp[2], 2);
-    const latest_blocks = await this.subnet.getLatestBlocks();
-    assert.equal(block2_resp[4], false);
-    assert.equal(latest_blocks["0"][0], block2_hash);
-    assert.equal(latest_blocks["1"][0], this.block1_hash);
-
-    const block2_resp2 = await this.subnet.getHeaderByNumber(2);
-    assert.equal(block2_resp2[0], block2_hash);
-    assert.equal(block2_resp2[1], 2);
-  });
-
-  it("Confirm A Received Block", async() => {
-    
-    var [block2, block2_encoded, block2_hash] = composeAndSignBlock(2, 2, 1, this.block1_hash, this.validators, 2, [], []);
-    var [block3, block3_encoded, block3_hash] = composeAndSignBlock(3, 3, 2, block2_hash, this.validators, 2, [], []);
-    var [block4, block4_encoded, block4_hash] = composeAndSignBlock(4, 4, 3, block3_hash, this.validators, 2, [], []);
-    var [block5, block5_encoded, block5_hash] = composeAndSignBlock(5, 5, 4, block4_hash, this.validators, 2, [], []);
-
-    await this.subnet.receiveHeader([block2_encoded, block3_encoded]); 
-    await this.subnet.receiveHeader([block4_encoded, block5_encoded]);
-
-    const block2_resp = await this.subnet.getHeader(block2_hash);
-    assert.equal(block2_resp[0], this.block1_hash);
-    assert.equal(block2_resp[1], 2);
-    assert.equal(block2_resp[2], 2);
-
-    const latest_blocks = await this.subnet.getLatestBlocks();
-    assert.equal(block2_resp[4], true);
-    assert.equal(latest_blocks["0"][0], block5_hash);
-    assert.equal(latest_blocks["1"][0], block2_hash);
-
-    const block2_resp2 = await this.subnet.getHeaderByNumber(2);
-    assert.equal(block2_resp2[0], block2_hash);
-    assert.equal(block2_resp2[1], 2);
-
-    const block3_resp = await this.subnet.getHeaderByNumber(3);
-    assert.equal(block3_resp[0], block3_hash);
-    assert.equal(block3_resp[1], 3);
-  });
-
-  it("Switch a Validator Set", async() => {
-    
-    var [block2, block2_encoded, block2_hash] = composeAndSignBlock(2, 2, 1, this.block1_hash, this.validators, 2, [], []);
-    var [block3, block3_encoded, block3_hash] = composeAndSignBlock(3, 3, 2, block2_hash, this.validators, 2, [], []);
-    var [block4, block4_encoded, block4_hash] = composeAndSignBlock(4, 4, 3, block3_hash, this.validators, 2, [], []);
-    var [block5, block5_encoded, block5_hash] = composeAndSignBlock(5, 5, 4, block4_hash, this.validators, 2, [], []);
-    let new_validators = createValidators(3);
-    var [block6, block6_encoded, block6_hash] = composeAndSignBlock(6, 6, 5, block5_hash, this.validators, 2, [], new_validators.map((val) => val.address));
-    var [block7, block7_encoded, block7_hash] = composeAndSignBlock(7, 7, 6, block6_hash, this.validators, 2, [], []);
-    var [block8, block8_encoded, block8_hash] = composeAndSignBlock(8, 8, 7, block7_hash, this.validators, 2, [], []);
-    var [block9, block9_encoded, block9_hash] = composeAndSignBlock(9, 9, 8, block8_hash, this.validators, 2, [], []);
-    var [block10, block10_encoded, block10_hash] = composeAndSignBlock(10, 10, 9, block9_hash, new_validators, 2, new_validators.map((val) => val.address), []);
-
-    await this.subnet.receiveHeader([block2_encoded, block3_encoded, block4_encoded]); 
-    await this.subnet.receiveHeader([block5_encoded, block6_encoded, block7_encoded]);
-    await this.subnet.receiveHeader([block8_encoded, block9_encoded, block10_encoded]);
-
-    const block7_resp = await this.subnet.getHeader(block7_hash);
-    assert.equal(block7_resp[0], block6_hash);
-    assert.equal(block7_resp[1], 7);
-    assert.equal(block7_resp[2], 7);
-
-    const latest_blocks = await this.subnet.getLatestBlocks();
-    assert.equal(block7_resp[4], true);
-    assert.equal(latest_blocks["0"][0], block10_hash);
-    assert.equal(latest_blocks["1"][0], block7_hash);
-
-    const block7_resp2 = await this.subnet.getHeaderByNumber(7);
-    assert.equal(block7_resp2[0], block7_hash);
-    assert.equal(block7_resp2[1], 7);
-
-    const block8_resp = await this.subnet.getHeaderByNumber(8);
-    assert.equal(block8_resp[0], block8_hash);
-    assert.equal(block8_resp[1], 8);
-  });
-
-  it("Switch a Validator Set in Special Case", async() => {
-    
-    var [block2, block2_encoded, block2_hash] = composeAndSignBlock(2, 2, 1, this.block1_hash, this.validators, 2, [], []);
-    var [block3, block3_encoded, block3_hash] = composeAndSignBlock(3, 3, 2, block2_hash, this.validators, 2, [], []);
-    var [block4, block4_encoded, block4_hash] = composeAndSignBlock(4, 4, 3, block3_hash, this.validators, 2, [], []);
-    var [block5, block5_encoded, block5_hash] = composeAndSignBlock(5, 5, 4, block4_hash, this.validators, 2, [], []);
-    let new_validators = createValidators(3);
-    var [block6, block6_encoded, block6_hash] = composeAndSignBlock(6, 9, 5, block5_hash, this.validators, 2, [], new_validators.map((val) => val.address));
-    var [block7, block7_encoded, block7_hash] = composeAndSignBlock(7, 10, 9, block6_hash, this.validators, 2, this.validators.map((val) => val.address), []);
-
-    await this.subnet.receiveHeader([block2_encoded, block3_encoded, block4_encoded]); 
-    await this.subnet.receiveHeader([block5_encoded, block6_encoded, block7_encoded]);
-
-    const latest_blocks = await this.subnet.getLatestBlocks();
-    assert.equal(latest_blocks["0"][0], block7_hash);
-    assert.equal(latest_blocks["1"][0], block2_hash);
-
-    const block5_resp = await this.subnet.getHeaderByNumber(5);
-    assert.equal(block5_resp[0], block5_hash);
-    assert.equal(block5_resp[1], 5);
-
-    const block2_resp = await this.subnet.getHeaderByNumber(2);
-    assert.equal(block2_resp[0], block2_hash);
-    assert.equal(block2_resp[1], 2);
-  });
-
-  // it("Lookup the transaction", async() => {
-  //   const raw_sigs = [];
-  //   const block1 = {
-  //     "number": 1,
-  //     "round_num": 0,
-  //     "parent_hash": this.genesis_hash,
-  //   };
-  //   const block1_hash = web3.utils.sha3(Buffer.from(
-  //     RLP.encode([
-  //       util.bigIntToUnpaddedBuffer(1),
-  //       util.bigIntToUnpaddedBuffer(0),
-  //       util.toBuffer(this.genesis_hash)
-  //   ])));
-  //   for (let i = 0; i < 3; i++) {
-  //     raw_sigs.push(
-  //       secp256k1.ecdsaSign(
-  //         hex2Arr(block1_hash.substring(2)),
-  //         hex2Arr(this.validators[i].privateKey.substring(2))
-  //     ));
-  //   }
-
-  //   const sigs = raw_sigs.map(x => {
-  //     var res = new Uint8Array(65);
-  //     res.set(x.signature, 0);
-  //     res.set([x.recid], 64);
-  //     return "0x"+Buffer.from(res).toString("hex");
-  //   });
-
-  //   await this.subnet.receiveHeader(block1, sigs);
-  //   const mainnet_num = await this.subnet.getMainnetBlockNumber(block1_hash);
-  //   const transactionCount = await web3.eth.getBlockTransactionCount(mainnet_num);
-  //   for (let i = 0; i < transactionCount; i++) {
-  //     let transaction = await web3.eth.getTransactionFromBlock(mainnet_num, i);
-  //     let decodeData = await this.decoder.decodeTransaction(transaction);
-  //     let block_hash = web3.utils.sha3(Buffer.from(
-  //       RLP.encode([
-  //         util.bigIntToUnpaddedBuffer(decodeData.arguments[0].value.value[0].value.value.asBN),
-  //         util.bigIntToUnpaddedBuffer(decodeData.arguments[0].value.value[1].value.value.asBN),
-  //         util.toBuffer(decodeData.arguments[0].value.value[2].value.value.asHex)
-  //     ])));
-  //     console.log(block_hash);
-  //   }
-  // });
-})
+});
